requirements.txt
streamlit
transformers
torch
sentencepiece
sentence-transformers
faiss-cpu
pypdf2
numpy
accelerate